+++ xarray/datatree_/datatree/mapping.py
@@ -189,10 +189,10 @@ def map_over_subtree(func: Callable) -> Callable:
             *args_as_tree_length_iterables,
             *list(kwargs_as_tree_length_iterables.values()),
         ):
-            node_args_as_datasets = [
+            node_args_as_datasetviews = [
                 a.ds if isinstance(a, DataTree) else a for a in all_node_args[:n_args]
             ]
-            node_kwargs_as_datasets = dict(
+            node_kwargs_as_datasetviews = dict(
                 zip(
                     [k for k in kwargs_as_tree_length_iterables.keys()],
                     [
@@ -204,7 +204,7 @@ def map_over_subtree(func: Callable) -> Callable:
 
             # Now we can call func on the data in this particular set of corresponding nodes
             results = (
-                func(*node_args_as_datasets, **node_kwargs_as_datasets)
+                func(*node_args_as_datasetviews, **node_kwargs_as_datasetviews)
                 if not node_of_first_tree.is_empty
                 else None
             )